requirements.txt

streamlit
requests
pydantic
pytest
supabase
python-dotenv
